Squashed commit of the following:
commit 008d138699316b279fbdf9b2eadb40d3fcc8962b
Author: Mateusz Cypcer <69849057+Motrixox@users.noreply.github.com>
Date:   Thu Mar 28 00:47:09 2024 +0100

    changed database provider to mongodb

+++ PuzzleBoost/SudokuWebService/ViewModels/SudokuBoardViewModel.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public int[][] Board { get; set; }
         public int[][] Solution { get; set; }
-        public float Difficulty { get; set; }
+        public double Difficulty { get; set; }
     }
 }
